# propagate inbounds