--- a/src/utils.jl
+++ b/src/utils.jl
@@ -28,8 +28,8 @@ swap row i and row j of v inplace, with f1, f2 factors applied on i and j (befor
 """
 function swaprows! end
 
-@inline function swaprows!(v::AbstractMatrix{T}, i::Int, j::Int, f1, f2) where T
-    @inbounds for c = 1:size(v, 2)
+Base.@propagate_inbounds function swaprows!(v::AbstractMatrix{T}, i::Int, j::Int, f1, f2) where T
+    for c = 1:size(v, 2)
         temp = v[i, c]
         v[i, c] = v[j, c]*f2
         v[j, c] = temp*f1
@@ -37,8 +37,8 @@ function swaprows! end
     v
 end
 
-@inline function swaprows!(v::AbstractMatrix{T}, i::Int, j::Int) where T
-    @inbounds for c = 1:size(v, 2)
+Base.@propagate_inbounds function swaprows!(v::AbstractMatrix{T}, i::Int, j::Int) where T
+    for c = 1:size(v, 2)
         temp = v[i, c]
         v[i, c] = v[j, c]
         v[j, c] = temp
@@ -53,8 +53,8 @@ swap col i and col j of v inplace, with f1, f2 factors applied on i and j (befor
 """
 function swapcols! end
 
-@inline function swapcols!(v::AbstractMatrix{T}, i::Int, j::Int, f1, f2) where T
-    @inbounds for c = 1:size(v, 1)
+Base.@propagate_inbounds function swapcols!(v::AbstractMatrix{T}, i::Int, j::Int, f1, f2) where T
+    for c = 1:size(v, 1)
         temp = v[c, i]
         v[c, i] = v[c, j]*f2
         v[c, j] = temp*f1
@@ -62,8 +62,8 @@ function swapcols! end
     v
 end
 
-@inline function swapcols!(v::AbstractMatrix{T}, i::Int, j::Int) where T
-    @inbounds for c = 1:size(v, 1)
+Base.@propagate_inbounds function swapcols!(v::AbstractMatrix{T}, i::Int, j::Int) where T
+    for c = 1:size(v, 1)
         temp = v[c, i]
         v[c, i] = v[c, j]
         v[c, j] = temp
@@ -71,16 +71,16 @@ end
     v
 end
 
-@inline swapcols!(v::AbstractVector, args...) = swaprows!(v, args...)
+Base.@propagate_inbounds swapcols!(v::AbstractVector, args...) = swaprows!(v, args...)
 
-@inline @inbounds function swaprows!(v::AbstractVector, i::Int, j::Int, f1, f2)
+Base.@propagate_inbounds function swaprows!(v::AbstractVector, i::Int, j::Int, f1, f2)
     temp = v[i]
     v[i] = v[j]*f2
     v[j] = temp*f1
     v
 end
 
-@inline @inbounds function swaprows!(v::AbstractVector, i::Int, j::Int)
+Base.@propagate_inbounds function swaprows!(v::AbstractVector, i::Int, j::Int)
     temp = v[i]
     v[i] = v[j]
     v[j] = temp
@@ -94,7 +94,7 @@ apply u1 on row i and row j of state inplace.
 """
 function u1rows! end
 
-@inline @inbounds function u1rows!(state::AbstractVector, i::Int, j::Int, a, b, c, d)
+Base.@propagate_inbounds function u1rows!(state::AbstractVector, i::Int, j::Int, a, b, c, d)
     w = state[i]
     v = state[j]
     state[i] = a*w+b*v
@@ -102,8 +102,8 @@ function u1rows! end
     state
 end
 
-@inline function u1rows!(state::AbstractMatrix, i::Int,j::Int, a, b, c, d)
-    @inbounds for col = 1:size(state, 2)
+Base.@propagate_inbounds function u1rows!(state::AbstractMatrix, i::Int,j::Int, a, b, c, d)
+    for col = 1:size(state, 2)
         w = state[i, col]
         v = state[j, col]
         state[i, col] = a*w+b*v
@@ -119,12 +119,16 @@ multiply row i of v by f inplace.
 """
 function mulrow! end
 
-@inline mulrow!(v::AbstractVector, i::Int, f) = (v[i] *= f; v)
-@inline function mulrow!(v::AbstractMatrix, i::Int, f)
-    @inbounds for j = 1:size(v, 2)
+Base.@propagate_inbounds function mulrow!(v::AbstractVector, i::Int, f)
+    v[i] *= f
+    return v
+end
+
+Base.@propagate_inbounds function mulrow!(v::AbstractMatrix, i::Int, f)
+    for j = 1:size(v, 2)
         v[i, j] *= f
     end
-    v
+    return v
 end
 
 """
@@ -134,13 +138,16 @@ multiply col i of v by f inplace.
 """
 function mulcol! end
 
+Base.@propagate_inbounds function mulcol!(v::AbstractVector, i::Int, f)
+    v[i] *= f
+    return v
+end
 
-@inline mulcol!(v::AbstractVector, i::Int, f) = (v[i] *= f; v)
-@inline function mulcol!(v::AbstractMatrix, j::Int, f)
-    @inbounds for i = 1:size(v, 1)
+Base.@propagate_inbounds function mulcol!(v::AbstractMatrix, j::Int, f)
+    for i = 1:size(v, 1)
         v[i, j] *= f
     end
-    v
+    return v
 end
 
 """
@@ -153,52 +160,52 @@ function matvec end
 matvec(x::AbstractMatrix) = size(x, 2) == 1 ? vec(x) : x
 matvec(x::AbstractVector) = x
 
-@inline function unrows!(state::AbstractVector, inds::AbstractVector, U::AbstractMatrix)
-    @inbounds state[inds] = U*state[inds]
-    state
+Base.@propagate_inbounds function unrows!(state::AbstractVector, inds::AbstractVector, U::AbstractMatrix)
+    state[inds] = U*state[inds]
+    return state
 end
 
-@inline function unrows!(state::AbstractMatrix, inds::AbstractVector, U::AbstractMatrix)
-    @inbounds for k in 1:size(state, 2)
+Base.@propagate_inbounds function unrows!(state::AbstractMatrix, inds::AbstractVector, U::AbstractMatrix)
+    for k in 1:size(state, 2)
         state[inds, k] = U*state[inds, k]
     end
-    state
+    return state
 end
 
 ############# boost unrows! for sparse matrices ################
 @inline unrows!(state::AbstractVector, inds::AbstractVector, U::IMatrix) = state
 
-@inline function unrows!(state::AbstractVector, inds::AbstractVector, U::SDDiagonal)
+Base.@propagate_inbounds function unrows!(state::AbstractVector, inds::AbstractVector, U::SDDiagonal)
     for i in 1:length(U.diag)
-        @inbounds state[inds[i]] *= U.diag[i]
+        state[inds[i]] *= U.diag[i]
     end
-    state
+    return state
 end
 
-@inline function unrows!(state::AbstractMatrix, inds::AbstractVector, U::SDDiagonal)
+Base.@propagate_inbounds function unrows!(state::AbstractMatrix, inds::AbstractVector, U::SDDiagonal)
     for j in 1:size(state, 2)
         for i in 1:length(U.diag)
-            @inbounds state[inds[i],j] *= U.diag[i]
+            state[inds[i],j] *= U.diag[i]
         end
     end
     state
 end
 
-@inline function unrows!(state::AbstractVector, inds::AbstractVector, U::SDPermMatrix)
-    @inbounds state[inds] = state[inds[U.perm]] .* U.vals
+Base.@propagate_inbounds function unrows!(state::AbstractVector, inds::AbstractVector, U::SDPermMatrix)
+    state[inds] = state[inds[U.perm]] .* U.vals
     state
 end
 
-@inline function unrows!(state::AbstractMatrix, inds::AbstractVector, U::SDPermMatrix)
-    @inbounds for k in 1:size(state, 2)
+Base.@propagate_inbounds function unrows!(state::AbstractMatrix, inds::AbstractVector, U::SDPermMatrix)
+    for k in 1:size(state, 2)
         state[inds, k] = state[inds[U.perm], k] .* U.vals
     end
     state
 end
 
-@inline function unrows!(state::AbstractVector, inds::AbstractVector, A::SDSparseMatrixCSC, work::AbstractVector)
+Base.@propagate_inbounds function unrows!(state::AbstractVector, inds::AbstractVector, A::SDSparseMatrixCSC, work::AbstractVector)
     work .= 0
-    @inbounds for col = 1:length(inds)
+    for col = 1:length(inds)
         xj = state[inds[col]]
         for j = A.colptr[col]:(A.colptr[col + 1] - 1)
             work[A.rowval[j]] += A.nzval[j]*xj
@@ -208,9 +215,9 @@ end
     state
 end
 
-@inline function unrows!(state::AbstractMatrix, inds::AbstractVector, A::SDSparseMatrixCSC, work::Matrix)
+Base.@propagate_inbounds function unrows!(state::AbstractMatrix, inds::AbstractVector, A::SDSparseMatrixCSC, work::Matrix)
     work .= 0
-    @inbounds for k = 1:size(state, 2)
+    for k = 1:size(state, 2)
         for col = 1:length(inds)
             xj = state[inds[col],k]
             for j = A.colptr[col]:(A.colptr[col + 1] - 1)