not a big problem, since julia know how to hoist the boundscheck sometimes now, but some of the instruct didn't correctly inbounds the callee

before:

```jl
julia> @benchmark instruct!(st, Val(:T), (3, )) setup=(st=rand(ComplexF64, 1<<15))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     17.568 μs (0.00% GC)
  median time:      19.893 μs (0.00% GC)
  mean time:        20.972 μs (0.00% GC)
  maximum time:     130.544 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

after

```jl
julia> @benchmark instruct!(st, Val(:T), (3, )) setup=(st=rand(ComplexF64, 1<<15))
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     14.798 μs (0.00% GC)
  median time:      17.689 μs (0.00% GC)
  mean time:        18.589 μs (0.00% GC)
  maximum time:     109.593 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```